+++ packages/hooks/src/index.ts
@@ -34,7 +34,8 @@ import useSet from './useSet';
 import usePersistFn from './usePersistFn';
 import useMap from './useMap';
 import useCreation from './useCreation';
-import { useDrag, useDrop } from './useDrop';
+import useDrop from './useDrop';
+import useDrag from './useDrag';
 import useMount from './useMount';
 import useTextSelection from './useTextSelection';
 import useCounter from './useCounter';
